backend/requirements.txt
fastapi
uvicorn
groq
sentence-transformers
faiss-cpu
youtube-transcript-api
numpy
python-dotenv
pydantic